Some more reworking
